Fixed bug in iOS
Apparently WebKit is more strict with lower and uppercase than Chrome: while Chrome considers a "p" and a "P" to be included in the string "People", iOS does not. The fix was to also lowercase the searchQuery, which I should have done regardless.
Also switched filter method from lodash to native javascript. 2022.1.8