Fixed search customer bug
Apparently WebKit is more strict with lower and uppercase than Chrome: while Chrome considers a "p" and a "P" to be included in the string "People", iOS does not. The fix was to also lowercase the searchQuery, which I should have done regardless.
Also switched the filter method from lodash to native javascript.